Move all printing functionality into Output
