+++ src/client/commands/stats.rs
@@ -2,7 +2,6 @@ use crate::client::globalsettings::GlobalSettings;
 use crate::rpc_call;
 use crate::transfer::connection::ClientConnection;
 use crate::transfer::messages::{FromClientMessage, StatsResponse, ToClientMessage};
-use cli_table::{print_stdout, Cell, Style, Table};
 
 pub async fn print_server_stats(
     gsettings: &GlobalSettings,
@@ -15,28 +14,6 @@ pub async fn print_server_stats(
     )
     .await?;
 
-    let rows = vec![
-        vec![
-            "Stream connections".cell().bold(true),
-            response.stream_stats.connections.join("\n").cell(),
-        ],
-        vec![
-            "Stream registrations".cell().bold(true),
-            response
-                .stream_stats
-                .registrations
-                .iter()
-                .map(|(job_id, path)| format!("{}: {}", job_id, path.display()))
-                .collect::<Vec<_>>()
-                .join("\n")
-                .cell(),
-        ],
-        vec![
-            "Open files".cell().bold(true),
-            response.stream_stats.files.join("\n").cell(),
-        ],
-    ];
-    let table = rows.table().color_choice(gsettings.color_policy());
-    assert!(print_stdout(table).is_ok());
+    gsettings.printer().print_server_stats(response);
     Ok(())
 }
