+++ src/client/globalsettings.rs
@@ -1,25 +1,24 @@
+use crate::client::output::outputs::Output;
 use std::path::{Path, PathBuf};
 
-use cli_table::ColorChoice;
-
 pub struct GlobalSettings {
-    color_policy: ColorChoice,
     server_dir: PathBuf,
+    printer: Box<dyn Output>,
 }
 
 impl GlobalSettings {
-    pub fn new(server_dir: PathBuf, color_policy: ColorChoice) -> Self {
+    pub fn new(server_dir: PathBuf, printer: Box<dyn Output>) -> Self {
         GlobalSettings {
-            color_policy,
             server_dir,
+            printer,
         }
-    }
-
-    pub fn color_policy(&self) -> ColorChoice {
-        self.color_policy
     }
 
     pub fn server_directory(&self) -> &Path {
         &self.server_dir
+    }
+
+    pub fn printer(&self) -> &dyn Output {
+        self.printer.as_ref()
     }
 }
