+++ src/client/mod.rs
@@ -1,7 +1,7 @@
 pub mod commands;
 pub mod globalsettings;
 pub mod job;
+pub mod output;
 pub mod resources;
 pub mod status;
 pub mod utils;
-pub mod worker;
