+++ src/worker/mod.rs
@@ -1,5 +1,4 @@
 pub mod hwdetect;
-pub mod output;
 pub mod parser;
 pub mod start;
 pub mod streamer;
